backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
alembic==1.13.1

# Data processing
pandas==2.1.4
numpy==1.25.2

# Machine learning
scikit-learn==1.3.2

# HTTP and API
httpx==0.25.2
requests==2.31.0

# Web scraping
beautifulsoup4==4.12.2

# Validation and serialization
pydantic==2.5.0

# Configuration and environment
python-dotenv==1.0.0
python-multipart==0.0.6

# Redis (optional)
redis==5.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1